+++ Command/Changes.md
@@ -1,5 +1,8 @@
 # Arbiter App - List of Changes
 
 ## v1.2 (in development)
+### New Features
+- Added edit company -feature (under `edit` command)
 ### Other Changes
+- Changed `edit` command default target from `product` to `company`
 - Local database now correctly defines UTF-8 charset